Beginning work on the jewelry overhaul

+++ src/main/java/mariculture/core/lib/EnchantSetting.java
@@ -4,7 +4,7 @@ public class EnchantSetting {
 	public static int JUMPS_PER;
 	public static double JUMP_FACTOR;
 	public static int SPEED_TICKS;
-	public static int SPEED_FACTOR;
+	public static double SPEED_FACTOR;
 	public static int TICK_REPAIR;
 	public static int RED_PEARL_DMG_CHANCE;
 }
